requirements.txt
scipy
h5py
pillow
imageio
nni
mmcv
tensorboard
pycocotools
tqdm
timm
fvcore